+++ Crates/orbital_procgeo/src/shapes/plane.rs
@@ -24,10 +24,10 @@ pub fn plane(size: Vector2<f32>, subdivisions: u32) -> MeshDescriptor {
 
     for j in 0..sw {
         for i in 0..sw {
-            let tl = (j * res + i) as u32;
-            let tr = (j * res + i + 1) as u32;
-            let bl = ((j + 1) * res + i) as u32;
-            let br = ((j + 1) * res + i + 1) as u32;
+            let tl = j * res + i;
+            let tr = j * res + i + 1;
+            let bl = (j + 1) * res + i;
+            let br = (j + 1) * res + i + 1;
             indices.extend_from_slice(&[tl, br, bl, tl, tr, br]);
         }
     }
